Remove donwcasting of int 64 to long

+++ main.c
@@ -19,9 +19,9 @@ int main(void) {
     int denoms[] = {1000000, 2000000, 7000000, 11000000, 20000000};
     
     for (int i = 0; i < 5; i++) {
-        long start_time = (long)get_absolute_time();
+        uint64_t start_time = get_absolute_time();
         avg(denoms[i]);
-        long end_time = (long)get_absolute_time();
+        uint64_t end_time = get_absolute_time();
     
         printf("Total take time in seconds for %d is: %ld\n", denoms[i], 
                 (end_time - start_time) / (1000000));
